add: for loops
I was going to try to prematurly optimize loop blocks but thankfully I backed off.
I will have to do this anyway when doing function as the arguments and interior block will have to be in the same scope.
Another option is to reintroduce a ListOfNodes ast node that does not create scope.