reformat: block scopes to not use call frames
The choice to use frames was horrible to being with but with closures coming up I had to change it. Now uses name mangling, though it would currently break if a closure captues a local in a block and is called after a second block uses the same variable.

fix: scope variable name resolution in block scopes