feat: add generics for builtins (List)
Removed the predetermined List types and added generics. I'm not super conviced by the implementation,
to make it work i had to move a lot of the logic into `Applies.solve` to make it work with overloads easily.
Now that I think about it, this might be the easier way to also handle user generics in the future too.